requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.2
python-dotenv>=1.0.0
pypdf>=3.17.1
beautifulsoup4>=4.12.2
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
spacy>=3.7.2
pytest>=7.4.0
jupyter>=1.0.0
python-docx>=1.0.0
tqdm>=4.66.1 